pyproject.toml
[project]
name = "isabel-wiki"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-generativeai>=0.8.6",
    "marker-pdf>=1.10.2",
    "networkx>=3.4.2",
    "nltk>=3.9.4",
    "psutil>=7.2.2",
    "pyyaml>=6.0",
    "requests>=2.33.1",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
]
